requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
beartype==0.17.0
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
decorator==5.1.1
decord==0.6.0
deepspeed==0.13.1
easydict==1.11
einops==0.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
executing==2.0.1
fairscale==0.4.13
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.1.3
google-auth==2.27.0
google-auth-oauthlib==1.0.0
grpcio==1.60.0
hjson==3.1.0
huggingface-hub==0.20.3
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
langcodes==3.3.0
lazy_loader==0.3
lightning-utilities==0.10.1
Markdown==3.5.2
MarkupSafe==2.1.4
matplotlib-inline==0.1.6
motion-vector-extractor==1.0.6
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
networkx==3.1
ninja==1.11.1.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
open-clip-torch==2.24.0
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pkgconfig==1.5.5
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.0
pydantic_core==2.16.1
Pygments==2.17.2
pynvml==11.5.0
pytorch-lightning==2.1.3
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rotary-embedding-torch==0.5.3
rsa==4.9
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
sentencepiece==0.1.99
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
thinc==8.2.2
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.12
tokenizers==0.15.1
tomli==2.0.1
torch==2.1.2
torchaudio==2.1.2
torchdiffeq==0.2.3
torchmetrics==1.3.0.post0
torchsde==0.2.6
torchvision==0.16.2
tqdm==4.66.1
traitlets==5.14.1
trampoline==0.1.2
transformers==4.37.2
triton==2.1.0
typer==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
Werkzeug==3.0.1
xformers==0.0.23.post1
yarl==1.9.4
zipp==3.17.0
